[Feat] Added Migration and updated database

+++ Core/CabukHemenSimdiIzle.Application/Features/Queries/MovieQueries/AddMovieRequest.cs
@@ -15,11 +15,11 @@ namespace CabukHemenSimdiIzle.Application.Features.Queries.MovieQueries
         public string Title { get; set; }
         public string Description { get; set; }
         public string? PictureUrl { get; set; }
-        public Int16 ReleaseDate { get; set; }
+        public short ReleaseDate { get; set; }
         public string Language { get; set; }
         public Genre Genre { get; set; }
         public decimal? ImdbRating { get; set; }
         public string AgeRestriction { get; set; }
-        public Int32 Duration { get; set; }
+        public short Duration { get; set; }
     }
 }
