+++ Core/CabukHemenSimdiIzle.Domain/Dtos/DirectorGetAllMoviesDto.cs
@@ -12,13 +12,13 @@ namespace CabukHemenSimdiIzle.Domain.Dtos
         public Guid Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public Int16 ReleaseDate { get; set; }
+        public short ReleaseDate { get; set; }
         public string Language { get; set; }
         public Genre Genre { get; set; }
         public decimal? ImdbRating { get; set; }
         public string AgeRestriction { get; set; }
 
-        public Int32 Duration { get; set; }
+        public short Duration { get; set; }
 
     }
 }
